# app/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ABRASEL REST API",
    "description": "Esta API REST é utilizada como curador dos dados de AFL da Abrasel em conjunto com os dados públicos da FRB"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http", "https"
  ],
  "paths": {
    "/api/v1/estabelecimentos/all/{page}/{pageSize}": {
      "get": {
        "description": "",
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}